refactor 'starter-packs' to 'resources'
