+++ src/App.js
@@ -8,7 +8,7 @@ import {
 	// Apply,
 	Home,
 	// Schedule,
-	StarterPacks
+	Resources
 } from "app/views";
 
 function App() {
@@ -19,7 +19,7 @@ function App() {
 				<Route path="/" element={<Home />} />
 				{/*<Route path="/apply" element={<Apply />} /> */}
 				{/* <Route path="/schedule" element={<Schedule />} /> */}
-				<Route path="/starter-packs" element={<StarterPacks />} />
+				<Route path="/resources" element={<Resources />} />
 				{/* redirect to Home if route not found */}
 				<Route path="*" element={<Navigate to="/" />} />
 			</Routes>
