+++ src/app/views/index.js
@@ -1,4 +1,4 @@
 export { default as Home } from "./home/Home";
 export { default as Schedule } from "./schedule/Schedule";
-export { default as StarterPacks } from "./starter-packs/StarterPacks";
+export { default as Resources } from "./resources/Resources";
 export { default as Apply } from "./apply/Apply";
